chore(motion): align naming of stories
