+++ packages/react-components/react-motion/stories/src/CreateMotionComponent/CreateMotionComponentFunctions.stories.tsx
@@ -1,22 +1,21 @@
 import {
-  createPresenceComponent,
+  createMotionComponent,
   Field,
   makeStyles,
   mergeClasses,
   type MotionImperativeRef,
-  type PresenceMotionFn,
+  motionTokens,
   Slider,
-  Switch,
   tokens,
 } from '@fluentui/react-components';
 import * as React from 'react';
 
-import description from './PresenceMotionFunctions.stories.md';
+import description from './CreateMotionComponentFunctions.stories.md';
 
 const useClasses = makeStyles({
   container: {
     display: 'grid',
-    gridTemplate: `"cardA cardB" "controls ." / 1fr 1fr`,
+    gridTemplate: `"card card" "controls ." / 1fr 1fr`,
     gap: '20px 10px',
   },
   card: {
@@ -24,13 +23,13 @@ const useClasses = makeStyles({
     flexDirection: 'column',
     alignItems: 'center',
     justifyContent: 'end',
+    gridArea: 'card',
 
     border: `${tokens.strokeWidthThicker} solid ${tokens.colorNeutralForeground3}`,
     borderRadius: tokens.borderRadiusMedium,
     boxShadow: tokens.shadow16,
     padding: '10px',
-
-    minHeight: '230px',
+    minHeight: '180px',
   },
   controls: {
     display: 'flex',
@@ -52,84 +51,50 @@ const useClasses = makeStyles({
     textWrap: 'nowrap',
   },
 
-  cardA: {
-    gridArea: 'cardA',
-  },
-  cardB: {
-    gridArea: 'cardB',
-  },
   item: {
     border: `${tokens.strokeWidthThicker} solid ${tokens.colorBrandBackground}`,
     padding: '8px',
-
     width: '300px',
+    overflow: 'hidden',
   },
-  description: {
-    fontFamily: tokens.fontFamilyMonospace,
-    borderRadius: tokens.borderRadiusMedium,
-    marginTop: '10px',
-    padding: '5px 10px',
-    backgroundColor: tokens.colorNeutralBackground1Pressed,
-  },
+  description: { margin: '5px' },
 });
 
-const collapseMotion: PresenceMotionFn = ({ element }) => {
-  const duration = 500;
-  const keyframes = [
-    { opacity: 0, maxHeight: '0px', overflow: 'hidden' },
-    { opacity: 1, maxHeight: `${element.scrollHeight}px`, overflow: 'hidden' },
-  ];
+const Grow = createMotionComponent(({ element }) => ({
+  duration: motionTokens.durationUltraSlow,
+  keyframes: [
+    { opacity: 0, maxHeight: `${element.scrollHeight / 2}px` },
+    { opacity: 1, maxHeight: `${element.scrollHeight}px` },
+    { opacity: 0, maxHeight: `${element.scrollHeight / 2}px` },
+  ],
+  iterations: Infinity,
+}));
 
-  return {
-    enter: { duration, keyframes },
-    exit: { duration, keyframes: [...keyframes].reverse() },
-  };
-};
-const Collapse = createPresenceComponent(collapseMotion);
-
-export const PresenceMotionFunctions = () => {
+export const CreateMotionComponentFunctions = () => {
   const classes = useClasses();
 
-  const motionInRef = React.useRef<MotionImperativeRef>();
-  const motionOutRef = React.useRef<MotionImperativeRef>();
-
-  const [playbackRate, setPlaybackRate] = React.useState<number>(30);
-  const [visible, setVisible] = React.useState<boolean>(true);
+  const motionRef = React.useRef<MotionImperativeRef>();
+  const [playbackRate, setPlaybackRate] = React.useState<number>(20);
 
   // Heads up!
   // This is optional and is intended solely to slow down the animations, making motions more visible in the examples.
   React.useEffect(() => {
-    motionInRef.current?.setPlaybackRate(playbackRate / 100);
-    motionOutRef.current?.setPlaybackRate(playbackRate / 100);
-  }, [playbackRate, visible]);
+    motionRef.current?.setPlaybackRate(playbackRate / 100);
+  }, [playbackRate]);
 
   return (
     <div className={classes.container}>
-      <div className={mergeClasses(classes.card, classes.cardA)}>
-        <Collapse imperativeRef={motionInRef} visible={visible}>
+      <div className={classes.card}>
+        <Grow imperativeRef={motionRef}>
           <div className={classes.item}>
             Lorem ipsum dolor sit amet, consectetuer adipiscing elit. Sed vel lectus. Donec odio tempus molestie,
             porttitor ut, iaculis quis, sem. Integer vulputate sem a nibh rutrum consequat. Etiam quis quam. Curabitur
             sagittis hendrerit ante. Duis ante orci, molestie vitae vehicula venenatis, tincidunt ac pede.
           </div>
-        </Collapse>
-        <div className={classes.description}>normal state</div>
-      </div>
-      <div className={mergeClasses(classes.card, classes.cardB)}>
-        <Collapse imperativeRef={motionOutRef} visible={!visible}>
-          <div className={classes.item}>
-            Lorem ipsum dolor sit amet, consectetuer adipiscing elit. Sed vel lectus. Donec odio tempus molestie,
-            porttitor ut, iaculis quis, sem. Integer vulputate sem a nibh rutrum consequat. Etiam quis quam. Curabitur
-            sagittis hendrerit ante. Duis ante orci, molestie vitae vehicula venenatis, tincidunt ac pede.
-          </div>
-        </Collapse>
-        <div className={classes.description}>reversed state</div>
+        </Grow>
       </div>
 
       <div className={classes.controls}>
-        <Field className={classes.field}>
-          <Switch label="Visible" checked={visible} onChange={() => setVisible(v => !v)} />
-        </Field>
         <Field
           className={mergeClasses(classes.field, classes.sliderField)}
           label={{
@@ -144,6 +109,7 @@ export const PresenceMotionFunctions = () => {
         >
           <Slider
             aria-valuetext={`Value is ${playbackRate}%`}
+            className={mergeClasses(classes.field, classes.sliderField)}
             value={playbackRate}
             onChange={(ev, data) => setPlaybackRate(data.value)}
             min={0}
@@ -156,7 +122,7 @@ export const PresenceMotionFunctions = () => {
   );
 };
 
-PresenceMotionFunctions.parameters = {
+CreateMotionComponentFunctions.parameters = {
   docs: {
     description: {
       story: description,
