--- packages/react-components/react-motion/stories/src/CreatePresenceComponent/PresenceMotionArrays.stories.tsx
@@ -1,157 +1,0 @@
-import {
-  createPresenceComponent,
-  Field,
-  makeStyles,
-  mergeClasses,
-  type MotionImperativeRef,
-  motionTokens,
-  Slider,
-  Switch,
-  tokens,
-} from '@fluentui/react-components';
-import * as React from 'react';
-
-import description from './PresenceMotionArrays.stories.md';
-
-const useClasses = makeStyles({
-  container: {
-    display: 'grid',
-    gridTemplate: `"card card" "controls ." / 1fr 1fr`,
-    gap: '20px 10px',
-  },
-  card: {
-    display: 'flex',
-    flexDirection: 'column',
-    alignItems: 'center',
-    justifyContent: 'end',
-    gridArea: 'card',
-
-    border: `${tokens.strokeWidthThicker} solid ${tokens.colorNeutralForeground3}`,
-    borderRadius: tokens.borderRadiusMedium,
-    boxShadow: tokens.shadow16,
-    padding: '10px',
-  },
-  controls: {
-    display: 'flex',
-    flexDirection: 'column',
-    gridArea: 'controls',
-
-    border: `${tokens.strokeWidthThicker} solid ${tokens.colorNeutralForeground3}`,
-    borderRadius: tokens.borderRadiusMedium,
-    boxShadow: tokens.shadow16,
-    padding: '10px',
-  },
-  field: {
-    flex: 1,
-  },
-  sliderField: {
-    gridTemplateColumns: 'min-content 1fr',
-  },
-  sliderLabel: {
-    textWrap: 'nowrap',
-  },
-
-  balloon: {
-    display: 'inline-block',
-    width: '80px',
-    height: '100px',
-    backgroundColor: tokens.colorBrandBackground,
-    borderRadius: '80%',
-    position: 'relative',
-    boxShadow: 'inset -10px -10px 0 rgba(0,0,0,0.07)',
-    margin: '20px 30px',
-    zIndex: 1,
-
-    '::before': {
-      content: "'▲'",
-      fontSize: '20px',
-      color: tokens.colorCompoundBrandBackgroundPressed,
-      display: 'block',
-      textAlign: 'center',
-      width: '100%',
-      position: 'absolute',
-      bottom: '-12px',
-      zIndex: -1,
-    },
-  },
-});
-
-const FastFadeSlowScale = createPresenceComponent({
-  enter: [
-    {
-      keyframes: [{ opacity: 0 }, { opacity: 1 }],
-      duration: motionTokens.durationFast,
-      easing: motionTokens.curveLinear,
-    },
-    {
-      keyframes: [{ transform: 'scale(0)' }, { transform: 'scale(1)' }],
-      duration: motionTokens.durationSlow,
-      easing: motionTokens.curveEasyEase,
-    },
-  ],
-  exit: {
-    keyframes: [{ opacity: 1 }, { opacity: 0 }],
-    duration: motionTokens.durationSlow,
-  },
-});
-
-export const PresenceMotionArrays = () => {
-  const classes = useClasses();
-
-  const motionRef = React.useRef<MotionImperativeRef>();
-  const ref = React.useRef<HTMLDivElement>(null);
-
-  const [visible, setVisible] = React.useState<boolean>(true);
-  const [playbackRate, setPlaybackRate] = React.useState<number>(30);
-
-  // Heads up!
-  // This is optional and is intended solely to slow down the animations, making motions more visible in the examples.
-  React.useEffect(() => {
-    motionRef.current?.setPlaybackRate(playbackRate / 100);
-  }, [playbackRate, visible]);
-
-  return (
-    <div className={classes.container}>
-      <div className={classes.card}>
-        <FastFadeSlowScale imperativeRef={motionRef} visible={visible}>
-          <div ref={ref} className={classes.balloon} />
-        </FastFadeSlowScale>
-      </div>
-
-      <div className={classes.controls}>
-        <Field className={classes.field}>
-          <Switch label="Visible" checked={visible} onChange={() => setVisible(v => !v)} />
-        </Field>
-        <Field
-          className={mergeClasses(classes.field, classes.sliderField)}
-          label={{
-            children: (
-              <>
-                <code>playbackRate</code>: {playbackRate}%
-              </>
-            ),
-            className: classes.sliderLabel,
-          }}
-          orientation="horizontal"
-        >
-          <Slider
-            aria-valuetext={`Value is ${playbackRate}%`}
-            value={playbackRate}
-            onChange={(ev, data) => setPlaybackRate(data.value)}
-            min={0}
-            max={100}
-            step={5}
-          />
-        </Field>
-      </div>
-    </div>
-  );
-};
-
-PresenceMotionArrays.parameters = {
-  docs: {
-    description: {
-      story: description,
-    },
-  },
-};
