+++ packages/react-components/react-motion/stories/src/CreatePresenceComponent/index.stories.ts
@@ -1,22 +1,23 @@
 import type { Meta } from '@storybook/react';
+
 import CreatePresenceComponentDescription from './CreatePresenceComponentDescription.md';
-import { CreatePresenceComponent } from './CreatePresenceComponent.stories';
+import { CreatePresenceComponentDefault } from './CreatePresenceComponentDefault.stories';
 
 export { CreatePresenceComponentDefault as Default } from './CreatePresenceComponentDefault.stories';
 
-export { CreatePresenceComponent as createPresenceComponent } from './CreatePresenceComponent.stories';
+export { CreatePresenceComponentFactory as createPresenceComponent } from './CreatePresenceComponentFactory.stories';
 
-export { PresenceAppear as appear } from './PresenceAppear.stories';
-export { PresenceUnmountOnExit as unmountOnExit } from './PresenceUnmountOnExit.stories';
-export { PresenceLifecycleCallbacks as LifecycleCallbacks } from './PresenceLifecycleCallbacks.stories';
+export { CreatePresenceComponentAppear as appear } from './CreatePresenceComponentAppear.stories';
+export { CreatePresenceComponentUnmountOnExit as unmountOnExit } from './CreatePresenceComponentUnmountOnExit.stories';
+export { CreatePresenceComponentLifecycleCallbacks as LifecycleCallbacks } from './CreatePresenceComponentLifecycleCallbacks.stories';
 
-export { PresenceMotionArrays as arrays } from './PresenceMotionArrays.stories';
-export { PresenceMotionFunctions as functions } from './PresenceMotionFunctions.stories';
-export { PresenceMotionFunctionParams as functionParams } from './PresenceMotionFunctionParams.stories';
+export { CreatePresenceComponentArrays as arrays } from './CreatePresenceComponentArrays.stories';
+export { CreatePresenceComponentFunctions as functions } from './CreatePresenceComponentFunctions.stories';
+export { CreatePresenceComponentFunctionParams as functionParams } from './CreatePresenceComponentFunctionParams.stories';
 
 export default {
   title: 'Motion/APIs/createPresenceComponent',
-  component: CreatePresenceComponent,
+  component: CreatePresenceComponentDefault,
   parameters: {
     docs: {
       description: {
